+++ tools/configure.py
@@ -2,8 +2,6 @@ import os
 import shutil
 from importlib import resources
 
-from tools.cli import Cmd
-
 namespace_kwargs = {"title": "configure commands", "description": "create config files in ~/.aec/"}
